Repaso

+++ Application.Common/Entity.cs
@@ -9,8 +9,8 @@ namespace Application.Common
 {
     public class Entity
     {
-        private long id;
+        private int id;
 
-        public long Id { get {  return this.id; } set  { this.id = value; } }
+        public int Id { get {  return this.id; } set  { this.id = value; } }
     }
 }
